refactor(migration): extract csv/id helpers for PO import
